Add count occurences of substring to bash_proto.sh
Add a function to count the occurences of a substring inside a string to
bash_proto.sh

The function essentially takes a copy of the original string, cuts out
all occurences of the substring with the // operator, then calculates
the difference in length of the original string and the reduced string,
which is exactly the length of the substring times the number it was cut
out, which is exactly the length of the substring (l) times the number it
was cut out (n). The result (l*n) is then divided by the length of the
substring (l): leading to (l*n)/l = n, which is the number of times the
substring occurs in the original string.